+++ 80col-mono/picoterm_core.h
@@ -39,7 +39,8 @@ void refresh_cursor();
 void clear_cursor();
 void print_cursor();
 void handle_new_character(unsigned char ch);
-unsigned char read_key_from_buffer(); // in main.c
+//unsigned char read_key_from_buffer(); // in main.c
+
 
 // for debugging purposes
 void print_ascii_value(unsigned char asc);